# apps/docs/api-reference/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Threadly Marketplace API",
    "description": "A comprehensive API for the Threadly marketplace platform, enabling users to buy and sell fashion items, manage orders, handle payments, and interact with the community.",
    "termsOfService": "https://threadly.app/terms",
    "contact": {
      "name": "Threadly Support",
      "url": "https://threadly.app/support",
      "email": "support@threadly.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.threadly.app",
      "description": "Production API"
    },
    {
      "url": "https://staging-api.threadly.app",
      "description": "Staging API"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/search/suggestions": {
      "get": {
        "tags": ["Search"],
        "summary": "Get search suggestions",
        "description": "Get autocomplete suggestions for search queries",
        "operationId": "getSearchSuggestions",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Partial search query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 50
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of suggestions",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10,
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "suggestions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from Clerk authentication"
      }
    }
  },
...
}